deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
altair
scikit-learn==1.2.2